requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
decorator==5.1.1
diff-match-patch==20230430
Django==4.2.16
django-bootstrap-v5==1.0.11
django-debug-toolbar==4.4.6
django-import-export==4.1.0
django-pandas==0.6.7
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
et-xmlfile==1.1.0
numpy==2.1.0
openpyxl==3.1.5
pandalchemy==0.1.19
pandas==2.2.2
pbr==6.1.0
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.3.18
sqlalchemy-migrate==0.13.0
sqlparse==0.5.1
tablib==3.5.0
tabulate==0.9.0
Tempita==0.5.2
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.1
